core: add more logging internals
